life_os/requirements.txt

# Life OS Dependencies
# Core packages for AI crew management and life optimization

# YAML processing for configuration files
PyYAML>=6.0.1

# Date and time utilities
python-dateutil>=2.8.2

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine learning and AI frameworks
scikit-learn>=1.3.0
transformers>=4.30.0

# HTTP requests and API integration
requests>=2.31.0
httpx>=0.24.0

# Environment and configuration management
python-dotenv>=1.0.0

# Logging and monitoring
loguru>=0.7.0

# Task scheduling and automation
schedule>=1.2.0
APScheduler>=3.10.0

# Data persistence
sqlite-utils>=3.34
sqlalchemy>=2.0.0

# Natural language processing
nltk>=3.8.1
spacy>=3.6.0

# Visualization and reporting
matplotlib>=3.7.0
plotly>=5.15.0

# Configuration validation
pydantic>=2.0.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Web interface (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0

# Async support
asyncio-mqtt>=0.13.0
aiofiles>=23.1.0
